# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wallet",
    "description": "wallet api socumention",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}/api",
      "description": "Developement",
      "variables": {
        "port": {
          "default": "59001"
        }
      }
    },
    {
      "url": "https://wallet.andyjjrt.cc/api",
      "description": "Stable"
    }
  ],
  "security": [
    {
      "login": []
    }
  ],
  "paths": {
...
    "/user/signup": {
      "post": {
        "summary": "Creates a user.",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "barcode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API"
                },
                "example": {
                  "success": true,
                  "message": "使用者註冊成功",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API"
                },
                "example": {
                  "success": false,
                  "message": "使用者註冊失敗 error:...",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/user/profile": {
      "get": {
        "summary": "Get user's profile.",
        "description": "get profile",
        "tags": [
          "user"
        ],
        "security": [
          {
            "login": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "time_choosen",
            "schema": {
              "type": "string"
            },
            "description": "time_choosen (not necessary)"
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                },
                "example": {
                  "success": true,
                  "message": "取得使用者資料成功",
                  "data": {
                    "user_id": "user_12345",
                    "username": "royshao",
                    "nickname": "roy",
                    "selected_wallet_id": "wallet_12345",
                    "barcode": "Yhfl3LUI",
                    "elec_invoice_agree": true,
                    "elec_invoice_agree_time": "2022-08-03 11:40:00",
                    "wallets": [],
                    "debtors": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/API"
                },
                "example": {
                  "success": false,
                  "message": "取得使用者資料失敗 error:...",
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "login": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)",
        "flows": {
          "implicit": {
            "authorizationUrl": "/oauth/google/login",
            "scopes": {}
          }
        }
      }
    },
...
    "schemas": {
      "API": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "user_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "nickname": {
                "type": "string"
              },
              "select_wallet_id": {
                "type": "string"
              },
              "barcode": {
                "type": "string"
              },
              "elec_invoice_agree": {
                "type": "boolean"
              },
              "elec_invoice_agree_time": {
                "type": "string"
              },
              "wallets": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "wallet_id": {
                      "type": "string"
                    },
                    "wallet_name": {
                      "type": "string"
                    },
                    "wallet_total": {
                      "type": "number"
                    },
                    "wallet_description": {
                      "type": "string"
                    },
                    "selected": {
                      "type": "boolean"
                    },
                    "record_num": {
                      "type": "number"
                    },
                    "wallet_barcode": {
                      "type": "string"
                    },
                    "records": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "record_id": {
                            "type": "string"
                          },
                          "wallet_record_tag_id": {
                            "type": "string"
                          },
                          "record_ordinary": {
                            "type": "number"
                          },
                          "record_name": {
                            "type": "string"
                          },
                          "record_description": {
                            "type": "string"
                          },
                          "record_amount": {
                            "type": "number"
                          },
                          "record_type": {
                            "type": "string"
                          },
                          "record_date": {
                            "type": "string"
                          },
                          "record_debtors": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "debtor_id": {
                                  "type": "string"
                                },
                                "debtor_name": {
                                  "type": "string"
                                },
                                "debtor_record_amount": {
                                  "type": "number"
                                }
                              }
                            }
                          },
                          "record_created_time": {
                            "type": "string"
                          },
                          "record_updated_time": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tag_id": {
                            "type": "string"
                          },
                          "tag_wallet_id": {
                            "type": "string"
                          },
                          "tag_ordinary": {
                            "type": "number"
                          },
                          "tag_name": {
                            "type": "string"
                          },
                          "tag_type": {
                            "type": "string"
                          },
                          "tag_created_time": {
                            "type": "string"
                          },
                          "tag_updated_time": {
                            "type": "string"
                          },
                          "tag_color": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "debtors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "debtor_id": {
                      "type": "string"
                    },
                    "debtor_user_id": {
                      "type": "string"
                    },
                    "debtor_name": {
                      "type": "string"
                    },
                    "debtor_amount": {
                      "type": "number"
                    },
                    "debtor_created_time": {
                      "type": "string"
                    },
                    "debtor_updated_time": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}